Refactor arithmetic.py into arithmetic/ module
